rename files for avoid a class in a file of the same name
Having a class `Source` in a file `source.py` is perfectly valid Python,
it confuses the Sphinx autosummary extention because the api doc files
end up with the same name.
There are workarounds and fixes online, but they all seem somewhat hacky
and brittle.
