+++ marxs/optics/baffles.py
