+++ marxs/optics/tests/test_all_optics.py
@@ -29,7 +29,7 @@ from ...base import _parse_position_keywords
 from ...design import (RowlandTorus, GratingArrayStructure,
                        LinearCCDArray, RowlandCircleArray,
                        RectangularGrid)
-from ..baffle import Baffle
+from ..baffles import Baffle
 from ..multiLayerMirror import MultiLayerMirror
 from ...simulator import Sequence
 from ...missions.chandra.hess import HETG
