+++ marxs/optics/tests/test_baffle.py
@@ -1,7 +1,7 @@
 # Licensed under GPL version 3 - see LICENSE.rst
 import numpy as np
 from astropy.table import Table
-from ..baffle import Baffle
+from ..baffles import Baffle
 
 
 def test_photons_through():
